Extract globals and types to their own files
